Remove ShadowDOM fallback
We no longer need a fallback for ShadowDOM since Edge will be updating to Chromium this year.  We can remove a lot of code here, speeding up the creation of components significantly. Please note that this changes the behaviour of isolated components so that they no longer attempt to be isolated.